продолжение рефакторинга
