+++ js/main.js
@@ -2,23 +2,14 @@ import { renderPictures } from './thumbnails.js';
 import { openPictureModal } from './big-picture.js';
 import { openForm } from './upload-form/form.js';
 import { getData } from './api.js';
+import { showAlert } from './util.js';
 
 const PHOTOS_COUNT = 25;
-const ALERT_SHOW_TIME = 5000;
 
 const pictureList = document.querySelector('.pictures');
 const uploadInput = document.querySelector('.img-upload__input');
 
 let photos = [];
-
-const showAlert = () => {
-  const alertContainer = document.querySelector('#data-error').content.querySelector('.data-error');
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-};
 
 // Отрисовка миниатюр
 getData()
@@ -27,9 +18,7 @@ getData()
     renderPictures(photos);
   })
   .catch(
-    () => {
-      showAlert();
-    }
+    () => showAlert()
   );
 
 const onPictureClick = (evt) => {
